Move access intelligence feature under bitwarden_license
