+++ bitwarden_license/bit-web/src/app/admin-console/organizations/organizations-routing.module.ts
@@ -76,6 +76,13 @@ const routes: Routes = [
           },
         ],
       },
+      {
+        path: "access-intelligence",
+        loadChildren: () =>
+          import("../../tools/access-intelligence/access-intelligence.module").then(
+            (m) => m.AccessIntelligenceModule,
+          ),
+      },
     ],
   },
 ];
