+++ bitwarden_license/bit-web/src/app/tools/access-intelligence/application-table.mock.ts
