+++ bitwarden_license/bit-web/src/app/tools/access-intelligence/critical-applications.component.ts
@@ -7,10 +7,9 @@ import { debounceTime, map } from "rxjs";
 import { I18nService } from "@bitwarden/common/platform/abstractions/i18n.service";
 import { SearchModule, TableDataSource, NoItemsModule, Icons } from "@bitwarden/components";
 import { CardComponent } from "@bitwarden/tools-card";
-
-import { HeaderModule } from "../../layouts/header/header.module";
-import { SharedModule } from "../../shared";
-import { PipesModule } from "../../vault/individual-vault/pipes/pipes.module";
+import { HeaderModule } from "@bitwarden/web-vault/app/layouts/header/header.module";
+import { SharedModule } from "@bitwarden/web-vault/app/shared";
+import { PipesModule } from "@bitwarden/web-vault/app/vault/individual-vault/pipes/pipes.module";
 
 import { applicationTableMockData } from "./application-table.mock";
 import { RiskInsightsTabType } from "./risk-insights.component";
