+++ bitwarden_license/bit-web/src/app/tools/access-intelligence/password-health-members-uri.component.spec.ts
@@ -1,49 +1,46 @@
 import { ComponentFixture, TestBed } from "@angular/core/testing";
 import { ActivatedRoute, convertToParamMap } from "@angular/router";
-import { mock } from "jest-mock-extended";
+import { mock, MockProxy } from "jest-mock-extended";
 import { of } from "rxjs";
 
-// eslint-disable-next-line no-restricted-imports
 import {
   MemberCipherDetailsApiService,
   PasswordHealthService,
 } from "@bitwarden/bit-common/tools/reports/risk-insights";
 import { ApiService } from "@bitwarden/common/abstractions/api.service";
 import { AuditService } from "@bitwarden/common/abstractions/audit.service";
+import { OrganizationService } from "@bitwarden/common/admin-console/abstractions/organization/organization.service.abstraction";
 import { I18nService } from "@bitwarden/common/platform/abstractions/i18n.service";
 import { PasswordStrengthServiceAbstraction } from "@bitwarden/common/tools/password-strength";
 import { CipherService } from "@bitwarden/common/vault/abstractions/cipher.service";
 import { TableModule } from "@bitwarden/components";
-import { TableBodyDirective } from "@bitwarden/components/src/table/table.component";
+import { LooseComponentsModule } from "@bitwarden/web-vault/app/shared";
+import { PipesModule } from "@bitwarden/web-vault/app/vault/individual-vault/pipes/pipes.module";
 
-import { LooseComponentsModule } from "../../shared";
-import { PipesModule } from "../../vault/individual-vault/pipes/pipes.module";
+import { PasswordHealthMembersURIComponent } from "./password-health-members-uri.component";
 
-import { PasswordHealthComponent } from "./password-health.component";
+describe("PasswordHealthMembersUriComponent", () => {
+  let component: PasswordHealthMembersURIComponent;
+  let fixture: ComponentFixture<PasswordHealthMembersURIComponent>;
+  let cipherServiceMock: MockProxy<CipherService>;
+  const passwordHealthServiceMock = mock<PasswordHealthService>();
 
-describe("PasswordHealthComponent", () => {
-  let component: PasswordHealthComponent;
-  let fixture: ComponentFixture<PasswordHealthComponent>;
   const activeRouteParams = convertToParamMap({ organizationId: "orgId" });
 
   beforeEach(async () => {
+    cipherServiceMock = mock<CipherService>();
     await TestBed.configureTestingModule({
-      imports: [PasswordHealthComponent, PipesModule, TableModule, LooseComponentsModule],
-      declarations: [TableBodyDirective],
+      imports: [PasswordHealthMembersURIComponent, PipesModule, TableModule, LooseComponentsModule],
       providers: [
-        { provide: CipherService, useValue: mock<CipherService>() },
+        { provide: CipherService, useValue: cipherServiceMock },
         { provide: I18nService, useValue: mock<I18nService>() },
         { provide: AuditService, useValue: mock<AuditService>() },
-        { provide: ApiService, useValue: mock<ApiService>() },
-        { provide: MemberCipherDetailsApiService, useValue: mock<MemberCipherDetailsApiService>() },
+        { provide: OrganizationService, useValue: mock<OrganizationService>() },
         {
           provide: PasswordStrengthServiceAbstraction,
           useValue: mock<PasswordStrengthServiceAbstraction>(),
         },
-        {
-          provide: PasswordHealthService,
-          useValue: mock<PasswordHealthService>(),
-        },
+        { provide: PasswordHealthService, useValue: passwordHealthServiceMock },
         {
           provide: ActivatedRoute,
           useValue: {
@@ -51,20 +48,24 @@ describe("PasswordHealthComponent", () => {
             url: of([]),
           },
         },
+        {
+          provide: MemberCipherDetailsApiService,
+          useValue: mock<MemberCipherDetailsApiService>(),
+        },
+        {
+          provide: ApiService,
+          useValue: mock<ApiService>(),
+        },
       ],
     }).compileComponents();
   });
 
   beforeEach(() => {
-    fixture = TestBed.createComponent(PasswordHealthComponent);
+    fixture = TestBed.createComponent(PasswordHealthMembersURIComponent);
     component = fixture.componentInstance;
-
-    fixture.detectChanges();
   });
 
   it("should initialize component", () => {
     expect(component).toBeTruthy();
   });
-
-  it("should call generateReport on init", () => {});
 });
