+++ bitwarden_license/bit-web/src/app/tools/access-intelligence/password-health.component.ts
@@ -1,10 +1,10 @@
+import { CommonModule } from "@angular/common";
 import { Component, DestroyRef, inject, OnInit } from "@angular/core";
 import { takeUntilDestroyed } from "@angular/core/rxjs-interop";
-import { FormControl, FormsModule } from "@angular/forms";
 import { ActivatedRoute } from "@angular/router";
-import { debounceTime, map } from "rxjs";
+import { map } from "rxjs";
 
-// eslint-disable-next-line no-restricted-imports
+import { JslibModule } from "@bitwarden/angular/jslib.module";
 import {
   MemberCipherDetailsApiService,
   PasswordHealthService,
@@ -15,42 +15,42 @@ import { PasswordStrengthServiceAbstraction } from "@bitwarden/common/tools/pass
 import { CipherService } from "@bitwarden/common/vault/abstractions/cipher.service";
 import { CipherView } from "@bitwarden/common/vault/models/view/cipher.view";
 import {
+  BadgeModule,
   BadgeVariant,
-  SearchModule,
+  ContainerComponent,
   TableDataSource,
   TableModule,
-  ToastService,
 } from "@bitwarden/components";
-
-import { HeaderModule } from "../../layouts/header/header.module";
-// eslint-disable-next-line no-restricted-imports
-import { SharedModule } from "../../shared";
-// eslint-disable-next-line no-restricted-imports
-import { PipesModule } from "../../vault/individual-vault/pipes/pipes.module";
+import { HeaderModule } from "@bitwarden/web-vault/app/layouts/header/header.module";
+import { OrganizationBadgeModule } from "@bitwarden/web-vault/app/vault/individual-vault/organization-badge/organization-badge.module";
+import { PipesModule } from "@bitwarden/web-vault/app/vault/individual-vault/pipes/pipes.module";
 
 @Component({
   standalone: true,
-  selector: "tools-password-health-members",
-  templateUrl: "password-health-members.component.html",
-  imports: [PipesModule, HeaderModule, SearchModule, FormsModule, SharedModule, TableModule],
+  selector: "tools-password-health",
+  templateUrl: "password-health.component.html",
+  imports: [
+    BadgeModule,
+    OrganizationBadgeModule,
+    CommonModule,
+    ContainerComponent,
+    PipesModule,
+    JslibModule,
+    HeaderModule,
+    TableModule,
+  ],
   providers: [PasswordHealthService, MemberCipherDetailsApiService],
 })
-export class PasswordHealthMembersComponent implements OnInit {
+export class PasswordHealthComponent implements OnInit {
   passwordStrengthMap = new Map<string, [string, BadgeVariant]>();
 
   passwordUseMap = new Map<string, number>();
 
   exposedPasswordMap = new Map<string, number>();
 
-  totalMembersMap = new Map<string, number>();
-
   dataSource = new TableDataSource<CipherView>();
 
   loading = true;
-
-  selectedIds: Set<number> = new Set<number>();
-
-  protected searchControl = new FormControl("", { nonNullable: true });
 
   private destroyRef = inject(DestroyRef);
 
@@ -60,13 +60,8 @@ export class PasswordHealthMembersComponent implements OnInit {
     protected auditService: AuditService,
     protected i18nService: I18nService,
     protected activatedRoute: ActivatedRoute,
-    protected toastService: ToastService,
     protected memberCipherDetailsApiService: MemberCipherDetailsApiService,
-  ) {
-    this.searchControl.valueChanges
-      .pipe(debounceTime(200), takeUntilDestroyed())
-      .subscribe((v) => (this.dataSource.filter = v));
-  }
+  ) {}
 
   ngOnInit() {
     this.activatedRoute.paramMap
@@ -92,39 +87,9 @@ export class PasswordHealthMembersComponent implements OnInit {
     await passwordHealthService.generateReport();
 
     this.dataSource.data = passwordHealthService.reportCiphers;
-
     this.exposedPasswordMap = passwordHealthService.exposedPasswordMap;
     this.passwordStrengthMap = passwordHealthService.passwordStrengthMap;
     this.passwordUseMap = passwordHealthService.passwordUseMap;
-    this.totalMembersMap = passwordHealthService.totalMembersMap;
     this.loading = false;
-  }
-
-  markAppsAsCritical = async () => {
-    // TODO: Send to API once implemented
-    return new Promise((resolve) => {
-      setTimeout(() => {
-        this.selectedIds.clear();
-        this.toastService.showToast({
-          variant: "success",
-          title: null,
-          message: this.i18nService.t("appsMarkedAsCritical"),
-        });
-        resolve(true);
-      }, 1000);
-    });
-  };
-
-  trackByFunction(_: number, item: CipherView) {
-    return item.id;
-  }
-
-  onCheckboxChange(id: number, event: Event) {
-    const isChecked = (event.target as HTMLInputElement).checked;
-    if (isChecked) {
-      this.selectedIds.add(id);
-    } else {
-      this.selectedIds.delete(id);
-    }
   }
 }
